# More efficient writing of floats and doubles