… zio-json/shared/src/main/scala/zio/json/internal/numbers.scala