Fix #639

Below are throughout results with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                            (size)   Mode  Cnt       Score      Error  Units
ArrayOfDoublesWriting.zioJson           128  thrpt    5  134091.361 ± 2887.965  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5  127156.328 ± 2426.482  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5  176211.663 ± 2145.087  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5  173410.157 ± 1271.356  ops/s
GeoJSONWriting.zioJson                  N/A  thrpt    5   13172.891 ±  323.040  ops/s
GeoJSONWriting.zioSchemaJson            N/A  thrpt    5   15327.015 ±  776.040  ops/s
```

#### After
```txt
Benchmark                            (size)   Mode  Cnt       Score     Error  Units
ArrayOfDoublesWriting.zioJson           128  thrpt    5  148101.179 ± 843.487  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5  151185.805 ± 915.708  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5  197963.383 ± 705.483  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5  195488.686 ± 471.393  ops/s
GeoJSONWriting.zioJson                  N/A  thrpt    5   17067.182 ± 742.874  ops/s
GeoJSONWriting.zioSchemaJson            N/A  thrpt    5   23091.151 ± 270.407  ops/s
```

TBD benchmark results for Scala.js